+++ src/components/index.jsx
@@ -1,5 +1,23 @@
+// dropdowns
+export { FilterDropdown, SortDropdown } from "./dropdowns";
+
+// label-editor
+export { LabelEditor } from "./label-editor";
+
+// navigation
 export { TopNav } from "./navigation/top-nav";
 export { SideNav } from "./navigation/side-nav";
-export { ColorPalette, NoteInputForm, NoteCard } from "./notes";
-export { SecondaryNav } from "./secondary-nav";
-export { LabelEditor } from "./label-editor";
+export { SecondaryNav } from "./navigation/secondary-nav";
+
+//notes
+export { ColorPalette } from "./notes/color-palette";
+export { NoteCard } from "./notes/note-card";
+export { NoteInputForm } from "./notes/note-input-form";
+export { NotesList } from "./notes/notes-list";
+export { PriorityList } from "./notes/priority-list";
+export { RichTextEditor } from "./notes/text-editor";
+
+// password input
+export { PasswordInput } from "./password-input";
+
+
