--- src/components/notes/index.jsx
@@ -1,3 +1,0 @@
-export { ColorPalette } from "../notes/color-palette";
-export { NoteInputForm } from "../notes/note-input-form";
-export { NoteCard } from "./note-card"
