+++ src/components/notes/text-editor/index.jsx
@@ -1,6 +1,7 @@
+import "./text-editor.css";
 import ReactQuill from "react-quill";
 import 'react-quill/dist/quill.snow.css';
-import "./text-editor.css";
+
   
 const modules = {
     toolbar: [
