+++ src/contexts/index.jsx
@@ -1,0 +1,4 @@
+// custom-hooks
+export { AuthProvider, useAuth } from "./auth-context";
+export { NoteProvider, useNote }  from "./note-context";
+export { ComponentProvider, useComponent } from "./component-context";
