--- src/pages/auth/index.jsx
@@ -1,2 +1,0 @@
-export { Login } from "./login";
-export { Signup } from "./signup";
