+++ src/pages/index.jsx
@@ -1,5 +1,9 @@
+// public pages
+export { Signup } from "./auth/signup";
+export { Login } from "./auth/login";
+
+// private pages
 export { Home } from "./home";
-export { Archive } from "./archive";
 export { Labels } from "./labels";
+export { Archive } from "./archive";
 export { Trash } from "./trash";
-export { Login, Signup} from "./auth";
