+++ src/reducers/index.jsx
@@ -1,0 +1,4 @@
+// reducers
+export { initialAuthValue, authReducerFunction } from "./auth-reducer";
+export { initialNotesData, noteReducer } from "./note-reducer";
+export { initialComponentValues, componentReducer } from "./component-reducer";
