+++ src/routes/index.jsx
@@ -1,0 +1,3 @@
+// routes
+export { PrivateRoutes } from "./private-routes";
+export { SiteRoutes } from "./site-routes";
