+++ src/routes/private-routes.jsx
@@ -1,6 +1,6 @@
 import "../../src/styles/main.css"
 import { Outlet, Navigate } from "react-router-dom";
-import { useAuth } from "../contexts/auth-context";
+import { useAuth } from "../contexts";
 import { SideNav } from "../components";
 
 const PrivateRoutes = () => {
