+++ src/services/archive-services/get-archive-service.jsx
@@ -1,0 +1,12 @@
+import axios from "axios";
+
+const getArchiveServices = (authToken) => {
+    const response = axios.get(
+        `/api/archives`,
+        { headers: { authorization: authToken } }
+    );
+
+    return response;
+}
+
+export { getArchiveServices };
